potential orbital lock issue fix
