+++ faclib/config.h
@@ -312,9 +312,6 @@ void         DecodePJ(int i, int *p, int *j);
 int          SpecSymbol(char *s, int kl);
 int          ConstructConfigName(char *s, int n, CONFIG *c);
 void         ListConfig(char *fn, int n, int *kg);
-int          IBisect(int k, int n, int *a);
-int          Bisect(void *p0, int n, int m, void *p,
-		    int (*comp)(const void *, const void *));
 int          InGroups(int kg, int ng, int *kgroup);
 int          InitConfig(void);
 int          ReinitConfig(int m);
