+++ faclib/interpolation.h
@@ -23,6 +23,7 @@
 #include <stdio.h>
 #include <string.h>
 #include "global.h"
+#include "array.h"
 #include "dbase.h"
 
 typedef struct _MOD_RECORD_ {
